+++ scripts/models/database.py
@@ -11,6 +11,8 @@ import influxdb_client
 import psycopg2
 from influxdb_client.client.write_api import SYNCHRONOUS
 
+from scripts.models.data import DeviceData, UnknownDeviceData
+
 
 class Database:
     """Base class of each other classes in this script."""
@@ -19,7 +21,7 @@ class Database:
 
 
 class PostgreSQL(Database):
-    """Class responsible for PostgreSQL database connection."""
+    """Class responsible for PostgreSQL database communication."""
 
     def __init__(self) -> None:
         """Initializes database and api connection."""
@@ -35,75 +37,49 @@ class PostgreSQL(Database):
         logging.debug(f"Connected to {self.__class__.__name__}")
 
     def __enter__(self) -> object:
-        """Return object itself."""
         return self
 
     def __exit__(self, exc_type, exc_value, exc_traceback) -> None:
         """Closes database and api connection."""
-        # if any exception ocurred during context process
+        # log error if any exception ocurred during context process
         if any((exc_type, exc_value, exc_traceback)):
             logging.error(exc_traceback)
-        # closes connection
+        # closes connection and api
         logging.debug(f"Closing {self.__class__.__name__} connection")
         self.api.close()
         self.client.close()
         logging.debug(f"{self.__class__.__name__} connection has been closed")
 
     @property
-    def known_devices_mac_addresses(self) -> Set[str]:
-        """Returns mac addresses of registered devices."""
+    def devices(self) -> List[DeviceData]:
+        """Returns list of registered devices."""
         try:
-            self.api.execute("SELECT mac_address FROM devices_device;")
-        except Exception:
-            logging.error(f"Unknown error occurred!\n{traceback.format_exc()}")
-            return {}
-        else:
-            return {mac_address[0] for mac_address in self.api.fetchall()}
-
-    @property
-    def unknown_devices_mac_addresses(self) -> Set[str]:
-        """Returns mac addresses of unregistered devices."""
-        try:
-            self.api.execute("SELECT mac_address FROM unknown_devices;")
-        except Exception:
-            logging.error(f"Unknown error occurred!\n{traceback.format_exc()}")
-            return {}
-        return {mac_address[0] for mac_address in self.api.fetchall()}
-
-    @property
-    def devices_air(self) -> List[dict]:
-        """Returns data of air devices"""
-        try:
-            # list that stores devices data
-            air_devices_data = []
-            # makes query to database
             self.api.execute(
                 """
-                    SELECT devices_device.name, devices_device.ip_address, devices_device.mac_address, rooms_room.name
-                    FROM devices_device
-                    INNER JOIN rooms_room
-                    ON devices_device.location_id = rooms_room.id
-                    WHERE devices_device.category = 'air';
-                    """
+                SELECT d.name, r.name, d.category, d.brand, d.mac_address, d.ip_address
+                FROM devices_device as d
+                INNER JOIN rooms_room as r
+                ON r.id = d.location_id;
+                """
             )
-            query_result = [
-                device_info for device_info in self.api.fetchall()
-            ]
-            # transforms query result to list of dictionaries
-            for row in query_result:
-                air_devices_data.append(
-                    {
-                        "name": row[0],
-                        "ip_address": row[1],
-                        "mac_address": row[2],
-                        "location": row[3],
-                    }
-                )
+            devices = [DeviceData(*row) for row in self.api.fetchall()]
         except Exception:
             logging.error(f"Unknown error occurred!\n{traceback.format_exc()}")
             return []
         else:
-            return air_devices_data
+            return devices
+
+    @property
+    def unknown_devices(self) -> List[UnknownDeviceData]:
+        """Returns list of unregistered devices."""
+        try:
+            self.api.execute("SELECT * FROM unknown_devices;")
+            unknown_devices = [UnknownDeviceData(*row) for row in self.api.fetchall()]
+        except Exception:
+            logging.error(f"Unknown error occurred!\n{traceback.format_exc()}")
+            return []
+        else:
+            return unknown_devices
 
 
 class InfluxDB(Database):